Large Shapefile Support
Support reading of shapefiles > 2GB in size.

Note: these shapefiles are technically not spec compliant

+++ modules/plugin/shapefile/src/main/java/org/geotools/data/shapefile/ShapefileIndexerBoundsHelper.java
@@ -160,7 +160,7 @@ class ShapefileIndexerBoundsHelper {
 
         @Override
         public void read(int recNumber, Envelope env) throws IOException {
-            int offset = shpIndex.getOffsetInBytes(recNumber);
+            long offset = shpIndex.getOffsetInBytes(recNumber);
             reader.goTo(offset);
             Record rec = reader.nextRecord();
             env.init(rec.minX, rec.maxX, rec.minY, rec.maxY);
@@ -168,7 +168,7 @@ class ShapefileIndexerBoundsHelper {
 
         @Override
         public void expandEnvelope(int recNumber, Envelope env) throws IOException {
-            int offset = shpIndex.getOffsetInBytes(recNumber);
+            long offset = shpIndex.getOffsetInBytes(recNumber);
             reader.goTo(offset);
             Record rec = reader.nextRecord();
             env.expandToInclude(rec.minX, rec.minY);
